fix(cwd_jail): apply CodeRabbit review (9 threads, 2 critical)
Windows (critical):
- grant_sid_access now reads the existing DACL via
  GetNamedSecurityInfoW and passes it into SetEntriesInAclW so the
  AppContainer ACE is *merged* into the existing ACL rather than
  replacing it. Previous code passed NULL for the old ACL, which on
  application would wipe owner / SYSTEM / Administrators perms and
  lock the user out of their own workspace.
- SidGuard now frees AppContainer SIDs via FreeSid (the documented
  free function for both CreateAppContainerProfile and
  DeriveAppContainerSidFromAppContainerName). LocalFree was the wrong
  matched-free; on some Windows builds it corrupts the heap.

Linux:
- Doc comment: Landlock is applied in the *child* (post-fork,
  pre-exec) via pre_exec, not the parent. Corrected the wording.
- Added AccessFs::Execute to the root + read_only rules. Without it
  the jailed child cannot execve binaries it found there — Landlock
  blocks execute on anything not explicitly granted.

Registry:
- create() now loops generate_id() until it produces an id not
  already in the index. Prevents the same-second time-based id from
  silently overwriting an existing record after a process restart.
- delete() now removes the directory from disk first and only then
  drops the in-memory record. A disk-removal failure no longer leaves
  the registry forgetting the record while the directory survives.
- spawn_in / spawn_in_with both route through a new jail_for() helper
  that validates record.dir canonicalizes under self.base before
  spawning. A corrupted index can no longer bypass the directory jail
  root.
- All registry mutation paths (open/create/rename/set_notes/delete/
  clear/persist) now emit log::debug / log::trace diagnostics with
  the `[cwd_jail]` prefix so sandbox + persistence issues are
  triagable.
- Inline test module split out into registry_test.rs (wired via
  #[cfg(test)] #[path = "registry_test.rs"] mod tests;) so
  registry.rs drops from ~656 to 367 lines, under the ~500-line
  module guideline.

Also added one new test (spawn_in_refuses_path_outside_base) covering
the new jail_for() containment guard, and renamed the index-corruption
test to align with the new delete behaviour (no rollback path —
delete refuses cleanly without ever removing).

All 49 unit + 7 integration tests still pass.
